work on platform_linux + standardization
- "fix" workflow
- standardize pin type to i16
- allow for pins to go unused
- fix arm linux builds
- complete linux gpio, i2c, and uart impls
- remove lgpio dependency
- fix host www reload support

+++ platform/adc.h
@@ -15,13 +15,13 @@
  * @param pins array of pins to setup for reading ADC signals
  * @param num_pins number of pins in `pins[]`
  */
-void adc_setup(const u32 pins[], u32 num_pins);
+void adc_setup(const i16 pins[], u32 num_pins);
 
 /**
  * Reads the ADC signal on `pin`.
  * @param pin pin to read ADC signal from
  * @return the voltage of the ADC signal in volts, or -1 if the pin is invalid
  */
-f64 adc_read_raw(u32 pin);
+f64 adc_read_raw(i16 pin);
 
 #endif // PLATFORM_SUPPORTS_ADC
