+++ platform/host/gpio.c
@@ -5,24 +5,24 @@
 
 #include "platform/gpio.h"
 
-void gpio_setup(u32 pin, PinMode mode) {
+void gpio_setup(i16 pin, PinMode mode) {
     return; // Not implemented
     (void)pin;
     (void)mode;
 }
 
-PinState gpio_state(u32 pin) {
+PinState gpio_state(i16 pin) {
     return STATE_LOW; // Not implemented
     (void)pin;
 }
 
-void gpio_set(u32 pin, PinState state) {
+void gpio_set(i16 pin, PinState state) {
     return; // Not implemented
     (void)pin;
     (void)state;
 }
 
-void gpio_toggle(u32 pin) {
+void gpio_toggle(i16 pin) {
     return; // Not implemented
     (void)pin;
 }
