+++ platform/pico/adc.c
@@ -31,14 +31,14 @@
 
 // clang-format on
 
-void adc_setup(const u32 pins[], u32 num_pins) {
+void adc_setup(const i16 pins[], u32 num_pins) {
     adc_init();
     for (u32 i = 0; i < num_pins; i++) {
         adc_gpio_init(pins[i]);
     }
 }
 
-f64 adc_read_raw(u32 pin) {
+f64 adc_read_raw(i16 pin) {
     // Force the SMPS into PWM mode while reading the ADC to reduce noise
     gpio_set(PIN_SMPS_MODE, SMPS_MODE_PWM);
     adc_select_input(PIN_TO_ADC_CHANNEL(pin));
