+++ src/ctrl/aircraft.h
@@ -23,11 +23,11 @@ typedef enum Mode {
 // Helper macros to determine if the user is currently inputting on the controls
 // If used, ensure to #include "io/receiver.h" and "sys/configuration.h"
 #define DEADBAND config.control[CONTROL_DEADBAND]
-#define ROLL_INPUT() (fabsf(receiver_get((u32)config.pins[PINS_INPUT_AIL], RECEIVER_MODE_DEGREE) - 90.f) > DEADBAND)
-#define PITCH_INPUT() (fabsf(receiver_get((u32)config.pins[PINS_INPUT_ELE], RECEIVER_MODE_DEGREE) - 90.f) > DEADBAND)
+#define ROLL_INPUT() (fabsf(receiver_get((i16)config.pins[PINS_INPUT_AIL], RECEIVER_MODE_DEGREE) - 90.f) > DEADBAND)
+#define PITCH_INPUT() (fabsf(receiver_get((i16)config.pins[PINS_INPUT_ELE], RECEIVER_MODE_DEGREE) - 90.f) > DEADBAND)
 #define YAW_INPUT()                                                                                                    \
     (receiver_has_rud() &&                                                                                             \
-     fabsf(receiver_get((u32)config.pins[PINS_INPUT_RUD], RECEIVER_MODE_DEGREE) - 90.f) > DEADBAND)
+     fabsf(receiver_get((i16)config.pins[PINS_INPUT_RUD], RECEIVER_MODE_DEGREE) - 90.f) > DEADBAND)
 // Throttle input usually isn't self-centering so it's more difficult to determine if there is input
 #define USER_INPUTTING() (ROLL_INPUT() || PITCH_INPUT() || YAW_INPUT())
