+++ src/io/servo.c
@@ -17,7 +17,7 @@
 
 #include "servo.h"
 
-void servo_enable(const u32 pins[], u32 num_pins) {
+void servo_enable(const i16 pins[], u32 num_pins) {
     printpre("servo", "setting up %lu servos", num_pins);
     if (!pwm_setup_write(pins, num_pins, config.general[GENERAL_SERVO_HZ])) {
         log_message(TYPE_FATAL, "Failed to enable PWM output!", 500, 0, true);
@@ -27,7 +27,7 @@ void servo_enable(const u32 pins[], u32 num_pins) {
     }
 }
 
-void servo_set(u32 pin, f32 degree) {
+void servo_set(i16 pin, f32 degree) {
 #if !SIMCONNECT
     // Ensure speed is within range 0-180deg
     degree = clampf(degree, 0.f, 180.f);
@@ -35,11 +35,11 @@ void servo_set(u32 pin, f32 degree) {
     pwm_write_raw(pin, mapf(degree, 0.f, 180.f, 500.f, 2500.f));
 #else
     SCFlightControl control;
-    if (pin == (u32)config.pins[PINS_SERVO_AIL]) {
+    if (pin == (i16)config.pins[PINS_SERVO_AIL]) {
         control = FCTRL_AIL;
-    } else if (pin == (u32)config.pins[PINS_SERVO_ELE]) {
+    } else if (pin == (i16)config.pins[PINS_SERVO_ELE]) {
         control = FCTRL_ELE;
-    } else if (pin == (u32)config.pins[PINS_SERVO_RUD]) {
+    } else if (pin == (i16)config.pins[PINS_SERVO_RUD]) {
         control = FCTRL_RUD;
     } else {
         return; // Not simulated
@@ -48,10 +48,10 @@ void servo_set(u32 pin, f32 degree) {
 #endif // !SIMCONNECT
 }
 
-void servo_test(u32 servos[], u32 num_servos, const f32 degrees[], u32 num_degrees, u32 pause_between_moves_ms) {
+void servo_test(i16 servos[], u32 num_servos, const f32 degrees[], u32 num_degrees, u32 pause_between_moves_ms) {
     for (u32 d = 0; d < num_degrees; d++) {
         for (u32 s = 0; s < num_servos; s++) {
-            if (servos[s] == (u32)config.pins[PINS_SERVO_BAY]) {
+            if (servos[s] == (i16)config.pins[PINS_SERVO_BAY]) {
                 // The drop servo will be set to the configured detents so as not to possibly break it
                 if (d < (num_degrees / 2)) {
                     servo_set(servos[s], config.control[CONTROL_DROP_DETENT_OPEN]);
@@ -66,23 +66,23 @@ void servo_test(u32 servos[], u32 num_servos, const f32 degrees[], u32 num_degre
     }
 }
 
-void servo_get_pins(u32 *servos, u32 *num_servos) {
+void servo_get_pins(i16 *servos, u32 *num_servos) {
     switch ((ControlMode)config.general[GENERAL_CONTROL_MODE]) {
         case CTRLMODE_3AXIS_ATHR:
         case CTRLMODE_3AXIS:
-            servos[0] = (u32)config.pins[PINS_SERVO_AIL];
-            servos[1] = (u32)config.pins[PINS_SERVO_ELE];
-            servos[2] = (u32)config.pins[PINS_SERVO_RUD];
-            servos[3] = (u32)config.pins[PINS_SERVO_BAY];
+            servos[0] = (i16)config.pins[PINS_SERVO_AIL];
+            servos[1] = (i16)config.pins[PINS_SERVO_ELE];
+            servos[2] = (i16)config.pins[PINS_SERVO_RUD];
+            servos[3] = (i16)config.pins[PINS_SERVO_BAY];
             *num_servos = 4;
             break;
         case CTRLMODE_2AXIS_ATHR:
         case CTRLMODE_2AXIS:
         case CTRLMODE_FLYINGWING_ATHR:
         case CTRLMODE_FLYINGWING:
-            servos[0] = (u32)config.pins[PINS_SERVO_AIL];
-            servos[1] = (u32)config.pins[PINS_SERVO_ELE];
-            servos[2] = (u32)config.pins[PINS_SERVO_BAY];
+            servos[0] = (i16)config.pins[PINS_SERVO_AIL];
+            servos[1] = (i16)config.pins[PINS_SERVO_ELE];
+            servos[2] = (i16)config.pins[PINS_SERVO_BAY];
             *num_servos = 3;
             break;
     }
